Separate out tools into standalone modules and specify extra deps (#176)
* remove extra `invoke_tool` on `AbstractOperator`
