+++ src/concrete-core/concrete/tools/meta.py
@@ -1,0 +1,57 @@
+import inspect
+
+TOOLS_REGISTRY: dict[str, "MetaTool"] = {}
+
+
+class MetaTool(type):
+    """
+    This metaclass enables dynamic string representation of class objects without needing to instantiate them.
+    """
+
+    def __new__(cls, name, bases, attrs):
+        method_info = []
+        for attr, value in attrs.items():
+            if attr.startswith("_"):
+                continue
+            if callable(value) or isinstance(value, (classmethod, staticmethod)):
+                func = value.__func__ if isinstance(value, (classmethod, staticmethod)) else value
+
+                docstring = func.__doc__.strip() if func.__doc__ else "No docstring provided"
+
+                signature = inspect.signature(func)
+                params = []
+                for param_name, param in signature.parameters.items():
+                    param_str = param_name
+                    if param.annotation != inspect.Parameter.empty:
+                        if hasattr(param.annotation, '__name__'):
+                            param_str += f": {param.annotation.__name__}"
+                        else:
+                            # Handle Union types
+                            param_str += f": {str(param.annotation)}"
+                    if param.default != inspect.Parameter.empty:
+                        param_str += f" = {param.default}"
+                    params.append(param_str)
+
+                return_str = ""
+                if signature.return_annotation != inspect.Signature.empty:
+                    if hasattr(signature.return_annotation, '__name__'):
+                        return_str = f" -> {signature.return_annotation.__name__}"
+                    else:
+                        # Handle Union types
+                        return_str = f" -> {str(signature.return_annotation)}"
+
+                method_signature = f"{attr}({', '.join(params)}){return_str}"
+                method_info.append(f"{method_signature}\n\t{docstring}")
+
+        attrs["_str_representation"] = f"Tool Name: {name}\nTool Methods:\n" + "\n".join(
+            f"   - {info}" for info in method_info
+        )
+        new_class = super().__new__(cls, name, bases, attrs)
+        TOOLS_REGISTRY[name] = new_class
+        return new_class
+
+    def __str__(cls):
+        return cls._str_representation
+
+    def __repr__(cls):
+        return str(cls)
